+++ src/components/ImageLoadModal.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import { onBeforeUnmount, onMounted, ref, watch } from 'vue';
+import { nextTick, onBeforeUnmount, onMounted, ref, watch } from 'vue';
 import { BvTriggerableEvent } from 'bootstrap-vue-next';
 import { useImageLoadStore } from '@/stores/imageLoadStore';
 import { ImageFile } from '@/imageFile';
 import { guessOrientation, type orientationGuessResult } from '@/util/orientationGuesser';
-import OrientationScroller from '@/components/OrientationScroller.vue';
 import { Orientation } from '@/enums/orientation';
+import { imageFromFile } from '@/util/imageFromFile';
 
 const imageLoadStore = useImageLoadStore();
 const disableHide = ref(true);
@@ -15,10 +15,39 @@ const imageScrollContainer = ref<HTMLDivElement | null>(null);
 const imageInput = ref<HTMLInputElement | null>(null);
 const orientations = ref<orientationGuessResult[]>([]);
 const screenHeight = ref(window.innerHeight);
+const processing = ref(false);
 
-watch(orientations, (newVal, oldVal) => {
-  console.log('orientations updated:', oldVal, '->', newVal);
+watch(orientations, (newVal) => {
+  nextTick().then(() => {
+    newVal.forEach((res) => {
+      const canvas = document.getElementById(res.image.sha + '-canvas') as HTMLCanvasElement;
+      if (!canvas) {
+        console.error('canvas not found');
+        return;
+      }
+      drawImageToCanvas(canvas, res.image);
+    });
+  });
 });
+
+async function drawImageToCanvas(canvas: HTMLCanvasElement, image: ImageFile) {
+  const context = canvas.getContext('2d');
+  if (!context) {
+    console.error('context not found');
+    return;
+  }
+  const img = new Image();
+  img.src = await imageFromFile(image.file);
+  img.onload = () => {
+    canvas.width = screenHeight.value * (img.width / img.height);
+    canvas.height = screenHeight.value;
+    context.clearRect(0, 0, canvas.width, canvas.height);
+    context.drawImage(img, 0, 0, img.width, img.height, 0, 0, canvas.width, canvas.height);
+  };
+  img.onerror = (error) => {
+    console.error('failed to load image', error);
+  };
+}
 
 async function loadImages() {
   if (!imageInput.value) {
@@ -56,7 +85,9 @@ async function handleImageLoad() {
         import_images.push(image);
       })
     );
+    processing.value = true;
     orientations.value = orientations.value.concat(await guessOrientation(import_images));
+    processing.value = false;
   }
 }
 
@@ -69,7 +100,7 @@ onMounted(() => {
 });
 
 const updateScreenHeight = () => {
-  screenHeight.value = window.innerHeight;
+  screenHeight.value = window.innerHeight / 4;
 };
 
 onMounted(() => {
@@ -87,14 +118,14 @@ onBeforeUnmount(() => {
     v-model="imageLoadStore.showLoadModal"
     title="Load images"
     @hide="handleHide"
-    scrollable
-    size="xl"
+    size="lg"
     hide-footer
+    scrollable
   >
     <input type="file" multiple accept="image/*" ref="imageInput" hidden />
-    <div class="d-flex flex-column justify-content-center align-items-center">
+    <div class="d-flex flex-column h-100">
       <!-- Top elements -->
-      <div class="flex-fill w-100 d-flex justify-content-center">
+      <div class="flex-fill h-100 w-100 d-flex justify-content-center">
         <div>
           <BButton @click="disableHide = !disableHide"> toggle disable hide </BButton>
         </div>
@@ -102,47 +133,50 @@ onBeforeUnmount(() => {
       <!-- the images to select -->
       <div
         v-if="orientations.length > 0"
-        class="flex-fill w-100 d-flex justify-content-center"
+        class="flex-fill h-100 w-100 d-flex justify-content-center"
         ref="imageScrollContainer"
       >
-        <div
-          class="flex-grow-1 overflow-y-auto d-flex flex-column justify-content-center align-items-center"
-        >
+        <div class="flex-grow-1 h-100 overflow-y-auto d-flex flex-column align-items-center">
           <h2>Left</h2>
-          <OrientationScroller
-            :image-clicked="imageClicked"
-            :orientations="orientations"
-            :orientation="Orientation.left"
-            :height="screenHeight / 3"
-          />
+          <div
+            v-for="res in orientations.filter((val) => val.orientation === Orientation.left)"
+            :key="res.image.sha"
+            :id="res.image.sha + '-container'"
+          >
+            <BButton @click="imageClicked(res.image)" variant="outline-dark" class="w-100">
+              <canvas :id="res.image.sha + '-canvas'" class="w-100 rounded border border-2" />
+            </BButton>
+          </div>
         </div>
-        <div
-          class="flex-grow-1 overflow-y-auto d-flex flex-column justify-content-center align-items-center"
-        >
+        <div class="flex-grow-1 h-100 overflow-y-auto d-flex flex-column align-items-center">
           <h2>Frontal</h2>
-          <OrientationScroller
-            :image-clicked="imageClicked"
-            :orientations="orientations"
-            :orientation="Orientation.front"
-            :height="screenHeight / 3"
-          />
+          <div
+            v-for="res in orientations.filter((val) => val.orientation === Orientation.front)"
+            :key="res.image.sha"
+            :id="res.image.sha + '-container'"
+          >
+            <BButton @click="imageClicked(res.image)" variant="outline-dark" class="w-100">
+              <canvas :id="res.image.sha + '-canvas'" class="w-100 rounded border border-2" />
+            </BButton>
+          </div>
         </div>
-        <div
-          class="flex-grow-1 overflow-y-auto d-flex flex-column justify-content-center align-items-center"
-        >
+        <div class="flex-grow-1 h-100 overflow-y-auto d-flex flex-column align-items-center">
           <h2>Right</h2>
-          <OrientationScroller
-            :image-clicked="imageClicked"
-            :orientations="orientations"
-            :orientation="Orientation.right"
-            :height="screenHeight / 3"
-          />
+          <div
+            v-for="res in orientations.filter((val) => val.orientation === Orientation.right)"
+            :key="res.image.sha"
+            :id="res.image.sha + '-container'"
+          >
+            <BButton @click="imageClicked(res.image)" variant="outline-dark" class="w-100">
+              <canvas :id="res.image.sha + '-canvas'" class="w-100 rounded border border-2" />
+            </BButton>
+          </div>
         </div>
       </div>
       <!-- Bottom row -->
       <div class="mt-auto w-100">
         <hr />
-        <div class="d-flex justify-content-center align-items-center">
+        <div class="d-flex">
           <BButton @click="loadImages" variant="outline-dark">Load</BButton>
           <BProgress :value="progress" :max="imageCount" show-value class="w-75 mx-4" />
           <BButton @click="nextImage" variant="primary">Next</BButton>
